+++ src/simmate/apps/ppdb/web_scraper.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from django.utils.timezone import make_aware
 from rich.progress import track
 
-from simmate.database.external_connectors import WebScraper
+from simmate.database.external_connectors.web_scrape import WebScraper
 from simmate.toolkit import Molecule