+++ src/simmate/database/external_connectors/__init__.py
@@ -1,3 +1,1 @@
 # -*- coding: utf-8 -*-
-
-from .web_scrape import WebScraper